Restructure config options into a separate module
